Fix weird bit shift issue
Something in the way the memory reading loop was working for some reason
caused bitshifts to always equate to zero !!??

In re-implementing as a 'for' loop the issue seems to have gone away.
I'm sure I've just been missing something critical here, but it looks
like we're back in a good place.